test(smartWallet): extract helpers from psm integration
